Move title/font stuff into title module and conditional compile
